Keep old data if add_data is called with a hashref.
This brings add_data($hashref)'s behaviour closer in line with
add_data($scalar) and add_data($arrayref), which append the new data to
whatever has already been added.

It's not perfect, though: you can add hashrefs to arrayrefs or to scalars,
but you can't add an arrayref or a scalar to a hashref.